Collect and expose sync success/failure metrics
This adds the collection and exposure of sync success and failures.

Two metrics are added:
* configmapsecrets_syncs is a counter for the total number of successful
  and failed syncs.
* configmapsecrets_internal_errors is a counter which counts errors by
  type of error. Also recorded is the kind, name and namespace of the
resource which caused the error, to aid in troubleshooting.

Intended usage of these metrics would be for monitoring purposes.
Alarming on elevated rates of internal errors, including the kind, name
and namespace information in the alert, greatly aids in diagnosing
errors.